requirements.txt
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.5.1
scikit-surprise>=1.1.4
lightfm>=1.17
fastapi>=0.111.1
uvicorn[standard]>=0.30.3
scipy>=1.13.1
